Fix acceptance of fractional parts of weights and height in UserParametersActivity
https://trello.com/c/rMPV1XTH

+++ app/src/main/java/korablique/recipecalculator/ui/userparameters/UserParametersActivity.java
@@ -406,9 +406,9 @@ public class UserParametersActivity extends BaseActivity {
         if (numberField.getText().toString().isEmpty()) {
             return false;
         }
-        int number;
+        float number;
         try {
-            number = Integer.parseInt(numberField.getText().toString());
+            number = Float.parseFloat(numberField.getText().toString());
         } catch (NumberFormatException e) {
             return false;
         }
